+++ lib/src/televerse/bot/bot.dart
@@ -31,6 +31,7 @@ class Bot<CTX extends Context> {
   /// Options to configure the logger.
   final LoggerOptions? _loggerOptions;
 
+  /// Acts as the dfault error handler
   void _defaultErrorHandler(BotError<CTX> err) {
     print("‼️ An error occurred while processing the update.");
     if (err.sourceIsMiddleware) {
@@ -48,6 +49,7 @@ class Bot<CTX extends Context> {
     _report();
   }
 
+  /// Prints the issue reporting statement
   static void _report() {
     print("---------------------");
     print(
@@ -162,7 +164,7 @@ class Bot<CTX extends Context> {
   }
 
   /// Handles the error in initial `getMe` call
-  Future<void> _thenHandleGetMeError(Object err, StackTrace st) async {
+  Future<void> _handleTheGetMeError(Object err, StackTrace st) async {
     if (err is DioException) {
       if (err.type == DioExceptionType.connectionTimeout ||
           err.type == DioExceptionType.receiveTimeout ||
@@ -244,7 +246,7 @@ class Bot<CTX extends Context> {
     try {
       await getMe();
     } catch (err, st) {
-      _thenHandleGetMeError(err, st);
+      _handleTheGetMeError(err, st);
     }
     // Set instance variable
     _instance = this;
@@ -403,6 +405,11 @@ class Bot<CTX extends Context> {
     }
   }
 
+  /// Additional infomration on the bot
+  ///
+  /// This space can be used for plugins or any other utilities to improve user experience
+  final Map<String, dynamic> additionalInfo = {};
+
   /// List of Handler Scopes
   final List<HandlerScope<CTX>> _handlerScopes = [];
 
@@ -485,7 +492,7 @@ class Bot<CTX extends Context> {
   /// bot.contextBuilder(MyCustomContext.new);
   /// ```
   ///
-  /// For detailed usage instructions and examples, visit the [Custom Context Documentation](https://televerse.web.app/doc/custom-context).
+  /// For detailed usage instructions and examples, visit the [Custom Context Documentation](https://televerse.xooniverse.com/advanced/custom-context.html).
   void contextBuilder(ContextConstructor<CTX> constructor) {
     _usingCustomContext = true;
     _contextConstructor = constructor;
@@ -550,7 +557,7 @@ class Bot<CTX extends Context> {
           "If you still encounter issues, make sure that your custom context class extends `Context` "
           "and its constructor matches the required parameters:\n\n"
           "  MyContext({required super.api, required super.me, required super.update});\n"
-          "📖 Check out the complete usage documentation here: https://televerse.web.app/doc/custom-context\n",
+          "📖 Check out the complete usage documentation here: https://televerse.xooniverse.com/advanced/custom-context.html \n",
         );
       }
 
@@ -578,7 +585,9 @@ class Bot<CTX extends Context> {
   /// This method simply initalizes the bot - performs the initial getMe request
   /// to fill the `bot.me` property.
   Future<void> init() async {
-    await _initializeBot();
+    if (!initialized) {
+      await _initializeBot();
+    }
   }
 
   /// Start polling or webhook listener for fetching updates.