+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.23.1
+
+- Exposed couple of more properties in Context
+- Added `Bot.additionalInfo`
+- Documentation website is up.
+
 # 1.23.0
 
 - 🤖 Bot API 7.8
@@ -71,7 +77,7 @@
 
 - 🥳 Support for Custom Context
 - Added detailed
-  [usage documentation for custom context here](https://televerse.web.app/doc/custom-context).
+  [usage documentation for custom context here](https://televerse.xooniverse.com/advanced/custom-context.html).
 - The custom context feature allows you to extend the base functionality of your
   bot by using your own custom context classes.
 - Added `Bot.contextBuilder` method